main/requirements.txt
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.5.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonformatter==0.3.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.1.3